Mod: `raw_mic_capture` callback re-implement to `RawMicCapture` with event dispatch
