+++ src/tests/js_manual/index.html
@@ -7,8 +7,8 @@
   <body>
     <ul>
       <li><a href="audio_connect/Try_AudioClientModule.html"  target="_blank">audio_connect/AudioClientModule</a></li>
-      <li><a href="audio_connect/Try_raw_mic_capture.html"    target="_blank">audio_connect/raw_mic_capture</a></li>
       <li><a href="audio_connect/Try_RawAudioStreamPlay.html" target="_blank">audio_connect/RawAudioStreamPlay</a></li>
+      <li><a href="audio_connect/Try_RawMicCapture.html"      target="_blank">audio_connect/RawMicCapture</a></li>
       <li><a href="Try_assertThrows_ecma.html"                target="_blank">assertThrows_ecma</a></li>
       <li><a href="Try_MixerUI.html"                          target="_blank">MixerUI</a></li>
     </ul>
